+++ Map/app/src/debug/res/values/google_maps_api.xml
@@ -4,15 +4,15 @@
 
     To get one, follow this link, follow the directions and press "Create" at the end:
 
-    https://console.developers.google.com/flows/enableapi?apiid=maps_android_backend&keyType=CLIENT_SIDE_ANDROID&r=CA:6B:52:32:A5:DC:7D:49:9B:A4:2B:05:FD:F8:57:D6:C6:33:D3:AF%3Bcom.example.map
+    https://console.developers.google.com/flows/enableapi?apiid=maps_android_backend&keyType=CLIENT_SIDE_ANDROID&r=63:43:B8:B4:85:4B:0A:25:81:72:22:81:03:DE:A0:92:E3:07:DB:61%3Bcom.example.deadcrumbs
 
     You can also add your credentials to an existing key, using these values:
 
     Package name:
-    com.example.map
+    com.example.deadcrumbs
 
     SHA-1 certificate fingerprint:
-    CA:6B:52:32:A5:DC:7D:49:9B:A4:2B:05:FD:F8:57:D6:C6:33:D3:AF
+    63:43:B8:B4:85:4B:0A:25:81:72:22:81:03:DE:A0:92:E3:07:DB:61
 
     Alternatively, follow the directions here:
     https://developers.google.com/maps/documentation/android/start#get-key
