+++ Map/app/src/main/AndroidManifest.xml
