+++ Map/app/src/main/res/values/strings.xml
@@ -1,4 +1,4 @@
 <resources>
-    <string name="app_name">Map</string>
+    <string name="app_name">Deadcrumbs</string>
     <string name="title_activity_maps">Map</string>
 </resources>
