+++ Map/settings.gradle
@@ -1,2 +1,2 @@
-rootProject.name='Map'
+rootProject.name='Deadcrumbs'
 include ':app'
